plugins/requirements.txt
boto3>=1.28.0
azure-identity>=1.12.0
azure-monitor-query>=1.0.0
azure-mgmt-rdbms>=10.2.0
requests>=2.28.0
paramiko>=2.10.0
PyYAML>=6.0.0
